move currentGuess state to PlayArea
