+++ src/PlayArea/PlayArea.js
@@ -1,3 +1,4 @@
+import { useState, useEffect } from 'react';
 import KeyboardArea from '../KeyboardArea/KeyboardArea';
 import PreviousGuesses from "./PreviousGuesses";
 import CurrentGuess from "./CurrentGuess";
@@ -9,6 +10,25 @@ const PlayArea = (props) => {
   //TEMPORARY DUMMY DATA
   //const previousGuesses = ['was', 'one', 'two', 'four', 'tear', 'blizzard'];
 
+  const [currentGuess, setCurrentGuess] = useState('');
+
+  //listen for keystrokes
+  useEffect(() => {
+    const handleKeyDown = (event) => {
+      //check if pressed key is a single letter
+      if(/^[a-zA-Z]{1}$/.test(event.key)) {
+        
+      }
+      if(event.key === 'Enter') {
+        setCurrentGuess(prevCurrentGuess => prevCurrentGuess.concat("a"));
+      }
+    };
+
+    window.addEventListener('keydown', handleKeyDown);
+    return () => {
+      window.removeEventListener('keydown', handleKeyDown);
+    };
+  }, []);
 
   return (
     <>
